+++ tests/errors_test.go
@@ -3,11 +3,11 @@ package tests
 import (
 	"bytes"
 	"encoding/json"
+	"github.com/andreagrandi/sentire/internal/cli/formatter"
+	"github.com/andreagrandi/sentire/internal/client"
+	"github.com/andreagrandi/sentire/internal/config"
 	"os"
 	"os/exec"
-	"sentire/internal/cli/formatter"
-	"sentire/internal/client"
-	"sentire/internal/config"
 	"strings"
 	"testing"
 )